+++ clients/derivatives-trading-coin-futures/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 11.0.0 - 2026-01-13
+
+### Added (1)
+
+#### REST API
+
+- `placeMultipleOrders()` (`POST /dapi/v1/batchOrders`)
+
+### Changed (3)
+
+- Update `@binance/common` library to version `2.2.0`.
+
+#### REST API
+
+- Update response schema for `keepaliveUserDataStream()`.
+- Update response schema for `autoCancelAllOpenOrders()`.
+
 ## 10.0.1 - 2025-12-19
 
 ### Changed (1)